Removed duplicated code from the LamportGeneratorThread
