+++ .gitignore
@@ -43,3 +43,6 @@ resources/*.md5
 
 # This library file is auto generated by browserify.
 src/assets/scripts/bitcoinjs-lib.js
+
+# Web worker libraries are auto generated
+src/assets/scripts/webworker
