+++ package.json
@@ -17,9 +17,11 @@
     "ionic:serve": "ionic-app-scripts serve",
     "release:android": "./scripts/build-android.sh",
     "upload:hockey-app": "./scripts/release-to-hockey-app.sh",
-    "postinstall": "browserify ./bitcoin-js-wrapper/wrapper.js --standalone bitcoinjs > ./src/assets/scripts/bitcoinjs-lib.js",
+    "postinstall": "npm run compile-bitcoin-js-wrapper && npm run compile-web-worker-lib",
     "validate-translations": "node ./scripts/translation_checker.js",
-    "validate-translations:ci": "node ./scripts/translation_checker.js --ci"
+    "validate-translations:ci": "node ./scripts/translation_checker.js --ci",
+    "compile-web-worker-lib": "tsc --p tsconfig.webworker.json && ./scripts/prepare-webworker-lib.sh",
+    "compile-bitcoin-js-wrapper": "browserify ./bitcoin-js-wrapper/wrapper.js --standalone bitcoinjs > ./src/assets/scripts/bitcoinjs-lib.js"
   },
   "dependencies": {
     "@angular/animations": "5.2.10",
