+++ scripts/prepare-webworker-lib.sh
@@ -1,0 +1,10 @@
+# Because we are lazy and do not want to use Webpack we instead use this script.
+# This script will strip certain lines from the compiled Javascript files which
+# would otherwise cause a runtime error.
+
+for file in src/assets/scripts/webworker/***/*
+do
+    sed -i '' '/export/d' $file
+    sed -i '' '/import/d' $file
+    sed -i '' '/var SHA1PRNG = prng.SHA1PRNG;/d' $file
+done
