+++ src/core/merkle/LamportGeneratorThread.spec.ts
@@ -1,9 +1,10 @@
-import { LamportGenerator } from "./LamportGenerator";
+import { LamportGeneratorThread, ILamportGeneratorThreadInput } from "./LamportGeneratorThread";
 
-describe("LamportGenerator", () => {
-    it("should return the correct amount of public keys", () => {
-        let generator = new LamportGenerator(
-            [
+describe("LamportGeneratorThread", () => {
+    it("call the callback function", () => {
+        let input: ILamportGeneratorThreadInput = {
+            startIndex: 1337,
+            seeds: [
                 new Int8Array([
                     1, 2, 3, 4, 5, 6, 7, 8, 9, 10
                 ]),
@@ -35,11 +36,13 @@ describe("LamportGenerator", () => {
                     1, 2, 3, 4, 5, 6, 7, 8, 9, 10
                 ])
             ],
-            10
-        );
-        
-        generator.fill();
+            count: 10
+        };
 
-        expect(generator.publicKeys.length).toBe(10);
+        let callbackSpy = jasmine.createSpy("callback", () => {});
+
+        LamportGeneratorThread(input, callbackSpy);
+
+        expect(callbackSpy).toHaveBeenCalled();
     });
 });
